Replace generic 'Get Started Now!' button with Mailchimp subscription form to close #51
Phase one of seperating out the three functions the catch-all 'Get Info' form is serving. This subscription form goes to our 'News and Events' list in Mailchimp, and will act as a conduit to talk to people wanting updates and information about hte school, but who are not yet ready to enroll for the Primer or apply for a longer class.